Update server.c
When a socket is closed, sometimes it enters a 'TIME_WAIT' state for a period of time to prevent the same port from being reused. The 'Address already in use' error caused by this situation may require waiting for a while. To solve this issue, I made the following change in my code: